requirements.txt
Flask==3.0.0
flask-wtf==1.2.1
WTForms==3.1.1
flask-login==0.6.3
argon2-cffi==23.1.0
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0
python-dotenv==1.0.0
# Two-Factor Authentication (TOTP)
pyotp==2.9.0
qrcode[pil]==7.4.2
# Security: Rate limiting and CORS
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
